chore: email
move email function to shared utils, send messages on app updates
