+++ lambda/.gitignore
@@ -1,5 +1,5 @@
 db.sqlite
 app/src/schemas
-app/src/helpers/customValidate.ts
+app/src/utils/customValidate.ts
 app/src/interfaces
 .env
