+++ lambda/actions/package.json
@@ -12,7 +12,6 @@
     "aws-sdk": "^2.655.0"
   },
   "dependencies": {
-    "axios": "^0.21.1",
     "octokit": "^1.2.0",
     "pg": "^8.6.0",
     "sequelize": "^6.6.4",
