+++ lambda/actions/src/interfaces.ts
@@ -8,17 +8,3 @@ export interface Response {
   isBase64Encoded?: boolean;
   body?: string;
 }
-
-export interface EmailOptions {
-  from?: string;
-  to: string;
-  body: string;
-  bodyType?: string;
-  cc?: string[];
-  bcc?: string[];
-  delayTS?: number;
-  encoding?: string;
-  priority?: 'normal' | 'low' | 'high';
-  subject?: string;
-  tag?: string;
-}
