+++ lambda/shared/interfaces.ts
@@ -32,3 +32,17 @@ export interface FormattedData {
   agreeWithTC?: boolean;
   publicAccess?: boolean;
 }
+
+export interface EmailOptions {
+  from?: string;
+  to: string;
+  body: string;
+  bodyType?: string;
+  cc?: string[];
+  bcc?: string[];
+  delayTS?: number;
+  encoding?: string;
+  priority?: 'normal' | 'low' | 'high';
+  subject?: string;
+  tag?: string;
+}
