+++ lambda/shared/utils/ches.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { EmailOptions } from './interfaces';
+import { EmailOptions } from '../interfaces';
 const url = require('url');
 
 const fetchChesToken = async (username, password) => {
